Split exec and SubProcess
